Avoid failures when default libvirt network does not exist
This is a follow-up change to [1]. Depending on operating system and
environment configuration, default libvirt network may not exist.
Right now, `Check for libvirt default network` task throws an error in
this case causing nova playbook to fail.

This change fixes that by instructing ansible to not throw an error
if `virsh net-list` fails with "Network not found: no network with
matching name" because it is acceptable to not have this network.

[1] https://review.opendev.org/c/openstack/openstack-ansible-os_nova/+/899768

Change-Id: If692bc94f421bc84ad9e6d43f548b68196a9e751