Ensure nova_device_spec is templated as JSON string
When the nova_device_spec variable is provided as either a string or a
mapping, ensure that it's templated as a JSON string.

Also handle either strings or mappings within nova_device_spec if it's
provided as a list.

Closes-Bug: 2057961
Change-Id: I7041a19547af580408ff704578cb8f12d37da1ae
(cherry picked from commit 501cf14342dd5519f064bca2ec999d204c00bc67)